refactor(article,draft): integrate draft domain to article domain
