+++ services/article/api/domain/src/main/java/nettee/blolet/article/domain/sub/DraftBlockStatus.java
@@ -1,4 +1,4 @@
-package nettee.draft.draftblock.domain.type;
+package nettee.blolet.article.domain.sub;
 
 public enum DraftBlockStatus {
     REMOVED,
